display_for_window and get_monitor_refresh_rate platform impls
